+++ CHANGELOG.md
@@ -12,7 +12,9 @@
 - **CI**: Support validation of YML typed manifests
 - Cache the arm check to prevent multiple calls of `uname` on *nix
 - **scoop-uninstall**: Warn user when he tries to uninstall dependency, which is still needed
-- **scoop-status**: Detect if applications, which were installed as dependency are still needed
+- **scoop-status**:
+    - Properly support yml typed manifests and dependencies
+    - Detect if applications, which were installed as dependency are still needed
 - **scoop-list**:
     - Fix detection of installed status for PowerShell 5
     - Show if application was installed as dependency